chore: extract fetching into a hook
